upgrade signup from sequelize to mysql2
