+++ my-app/server/server.ts
@@ -40,21 +40,6 @@ app.use(
 app.use(passport.initialize());
 app.use(passport.session());
 
-const connection = mysql.createConnection({
-  host: process.env.HOST,
-  user: process.env.USERNAME1,
-  password: process.env.PASSWORD,
-  database: process.env.SIGNUPDB,
-});
-
-connection.connect((err: Error) => {
-  if (err) {
-    console.log(err);
-  } else {
-    console.log("db-mysql 연결성공");
-  }
-});
-
 sequelize
   .sync({ alert: true })
   .then(() => {
@@ -83,5 +68,3 @@ app.use((err: any, req: any, res: any, next: any) => {
 app.listen(app.get("port"), function () {
   console.log("listening 8081");
 });
-
-export {};
